+++ backend/pkg/kafka/delete_topics.go
@@ -1,0 +1,20 @@
+package kafka
+
+import (
+	"context"
+	"fmt"
+	"github.com/twmb/franz-go/pkg/kmsg"
+)
+
+func (s *Service) DeleteTopics(ctx context.Context, topicNames []string) (*kmsg.DeleteTopicsResponse, error) {
+	req := kmsg.NewDeleteTopicsRequest()
+	req.TopicNames = topicNames
+	req.TimeoutMillis = 30 * 1000 // 30s
+
+	res, err := req.RequestWith(ctx, s.KafkaClient)
+	if err != nil {
+		return nil, fmt.Errorf("failed to delete topics: %w", err)
+	}
+
+	return res, nil
+}